Move network messages to separate file
