+++ Assets/Scripts/Networking/NetworkLogic.cs
@@ -1,17 +1,11 @@
 ﻿using Mirror;
 using UnityEngine;
 using Open.Nat;
+using NetworkMessages;
 
 public class NetworkLogic : MonoBehaviour {
     Mapping port = new Mapping(Protocol.Tcp, 7777, 7777);
     NatDevice device;
-
-    public class MoveMessage : MessageBase {
-        public int originI;
-        public int originJ;
-        public int destI;
-        public int destJ;
-    }
 
     private async void Awake() {
         NatDiscoverer discoverer = new NatDiscoverer();
